+++ src/app/components/editors/code-editor/helper/html-linter-message.interface.ts
@@ -1,0 +1,8 @@
+import { HtmlLinterRule } from './html-linter-rule.enum';
+
+export interface HtmlLinterMessageInterface
+{
+    line:number;
+    col:number;
+    rule:HtmlLinterRule;
+}