+++ src/app/components/example-collection.ts
@@ -1,0 +1,90 @@
+import { Type } from '@angular/core';
+import { TerraButtonComponentExample } from './buttons/button/example/terra-button.component.example';
+import { TerraAlertComponentExample } from './alert/example/terra-alert.component.example';
+import { TerraCardComponentExample } from './layouts/card/example/terra-card.component.example';
+import { TerraBaseToolbarComponentExample } from './toolbar/base-toolbar/example/terra-base-toolbar.component.example';
+import { TerraOverlayComponentExample } from './layouts/overlay/example/terra-overlay.component.example';
+import { TerraInfoboxComponentExample } from './layouts/info-box/example/terra-info-box.component.example';
+import { TerraPortletComponentExample } from './layouts/portlet/example/terra-portlet.component.example';
+import { TerraColorPickerComponentExample } from './forms/input/color-picker/example/terra-color-picker.component.example';
+import { TerraDatePickerComponentExample } from './forms/input/date-picker/example/terra-date-picker.component.example';
+import { TerraDoubleInputComponentExample } from './forms/input/double-input/example/terra-double-input.component.example';
+import { TerraNumberInputComponentExample } from './forms/input/number-input/example/terra-number-input.component.example';
+import { TerraTextAreaInputComponentExample } from './forms/input/text-area-input/example/terra-text-area-input.component.example';
+import { TerraTextInputComponentExample } from './forms/input/text-input/example/terra-text-input.component.example';
+import { TerraCheckboxComponentExample } from './forms/checkbox/example/terra-checkbox.component.example';
+import { TerraMultiSelectBoxComponentExample } from './forms/multi-select-box/example/terra-multi-select-box.component.example';
+import { TerraMultiCheckBoxComponentExample } from './forms/multi-check-box/example/terra-multi-check-box.component.example';
+import { TerraRadioButtonComponentExample } from './forms/radio-button/example/terra-radio-button.component.example';
+import { TerraSelectBoxComponentExample } from './forms/select-box/example/terra-select-box.component.example';
+import { TerraSliderComponentExample } from './forms/slider/example/terra-slider.component.example';
+import { TerraSuggestionBoxComponentExample } from './forms/suggestion-box/example/terra-suggestion-box.component.example';
+import { TerraDataTableComponentExample } from './tables/data-table/example/terra-data-table.component.example';
+import { TerraTagComponentExample } from './layouts/tag/example/terra-tag.component.example';
+import { TerraFileInputComponentExample } from './forms/input/file-input/example/terra-file-input.component.example';
+import { TerraTreeComponentExample } from './tree/example/terra-tree.component.example';
+import { TerraCheckboxTreeComponentExample } from './tree/checkbox-tree/example/terra-checkbox-tree.component.example';
+import { TerraToggleComponentExample } from './buttons/toggle/example/terra-toggle.component.example';
+import { TerraSyntaxEditorComponentExample } from './editors/syntax-editor/example/terra-syntax-editor.component.example';
+import { TerraSimpleTableComponentExample } from './tables/simple/example/terra-table-simple.component.example';
+import { TerraPagerComponentExample } from './pager/example/terra-pager.component.example';
+import { TerraTaglistComponentExample } from './layouts/taglist/example/terra-taglist.component.example';
+import { TerraNavigatorComponentExample } from './navigator/example/terra-navigator.component.example';
+import { TerraNoResultNoticeComponentExample } from './no-result/example/terra-no-result-notice.component.example';
+import { TerraButtonWithOptionsComponentExample } from './buttons/button-with-options/example/terra-button-with-options.component.example';
+import { TerraInputComponentExample } from './forms/input/example/terra-input.component.example';
+import { TerraFilterComponentExample } from './filter/example/terra-filter.component.example';
+import { TerraDataTableContextMenuComponentExample } from './tables/data-table/context-menu/example/terra-data-table-context-menu.component.example';
+import { TerraNoteEditorComponentExample } from './editors/note-editor/example/terra-note-editor.component.example';
+import { TerraCodeEditorComponentExample } from './editors/code-editor/example/terra-code-editor.component.example';
+import { TerraNoteComponentExample } from './note/example/terra-note.component.example';
+import { TerraNodeTreeComponentExample } from './tree/node-tree/example/terra-node-tree.component.example';
+import { TerraTimePickerComponentExample } from './forms/input/time-picker/example/terra-time-picker.component.example';
+import { TerraNestedDataPickerComponentExample } from './nested-data-picker/example/terra-nested-data-picker.component.example';
+import { TerraTagSelectComponentExample } from './forms/tag-select/example/terra-tag-select.component.example';
+
+export const examples:Array<Type<any>> = [
+    TerraButtonComponentExample,
+    TerraAlertComponentExample,
+    TerraCardComponentExample,
+    TerraBaseToolbarComponentExample,
+    TerraOverlayComponentExample,
+    TerraInfoboxComponentExample,
+    TerraPortletComponentExample,
+    TerraColorPickerComponentExample,
+    TerraDatePickerComponentExample,
+    TerraDoubleInputComponentExample,
+    TerraNumberInputComponentExample,
+    TerraTextAreaInputComponentExample,
+    TerraTextInputComponentExample,
+    TerraCheckboxComponentExample,
+    TerraMultiSelectBoxComponentExample,
+    TerraMultiCheckBoxComponentExample,
+    TerraRadioButtonComponentExample,
+    TerraSelectBoxComponentExample,
+    TerraSliderComponentExample,
+    TerraSuggestionBoxComponentExample,
+    TerraDataTableComponentExample,
+    TerraTagComponentExample,
+    TerraFileInputComponentExample,
+    TerraTreeComponentExample,
+    TerraCheckboxTreeComponentExample,
+    TerraToggleComponentExample,
+    TerraSyntaxEditorComponentExample,
+    TerraSimpleTableComponentExample,
+    TerraPagerComponentExample,
+    TerraTaglistComponentExample,
+    TerraNavigatorComponentExample,
+    TerraNoResultNoticeComponentExample,
+    TerraButtonWithOptionsComponentExample,
+    TerraInputComponentExample,
+    TerraFilterComponentExample,
+    TerraDataTableContextMenuComponentExample,
+    TerraNoteEditorComponentExample,
+    TerraCodeEditorComponentExample,
+    TerraNoteComponentExample,
+    TerraNodeTreeComponentExample,
+    TerraTimePickerComponentExample,
+    TerraNestedDataPickerComponentExample,
+    TerraTagSelectComponentExample
+];