+++ src/app/components/forms/input/double-input/terra-double-input.component.ts
@@ -7,6 +7,7 @@ import {
 import { TerraInputComponent } from '../terra-input.component';
 import { NG_VALUE_ACCESSOR } from '@angular/forms';
 import { TerraRegex } from '../../../../helpers/regex/terra-regex';
+import { DefaultLocale } from 'angular-l10n';
 
 let nextId:number = 0;
 
@@ -37,12 +38,15 @@ export class TerraDoubleInputComponent extends TerraInputComponent implements On
     @Input()
     public inputDecimalCount:number = 2;
 
-    private _step:number;
+    @DefaultLocale()
+    protected locale:string;
+
+    protected step:number;
 
     /**
      * @description a unique string identifier for the specific input instance.
      */
-    private _id:string;
+    protected id:string;
 
     /**
      * @deprecated
@@ -61,13 +65,13 @@ export class TerraDoubleInputComponent extends TerraInputComponent implements On
         super(TerraRegex.DOUBLE);
 
         // generate the id of the input instance
-        this._id = `double-input_#${nextId++}`;
+        this.id = `double-input_#${nextId++}`;
     }
 
     public ngOnInit():void
     {
         this.regex = TerraRegex.getDouble(this.inputDecimalCount);
-        this._step = 1 / (Math.pow(10, this.inputDecimalCount));
+        this.step = 1 / (Math.pow(10, this.inputDecimalCount));
     }
 
     /**
@@ -77,7 +81,7 @@ export class TerraDoubleInputComponent extends TerraInputComponent implements On
     {
         setTimeout(() =>
         {
-            let input:HTMLInputElement = <HTMLInputElement> document.getElementById(this._id);
+            let input:HTMLInputElement = <HTMLInputElement> document.getElementById(this.id);
             input.focus();
         });
     }
@@ -89,7 +93,7 @@ export class TerraDoubleInputComponent extends TerraInputComponent implements On
     {
         setTimeout(() =>
         {
-            let input:HTMLInputElement = <HTMLInputElement> document.getElementById(this._id);
+            let input:HTMLInputElement = <HTMLInputElement> document.getElementById(this.id);
             input.select();
         });
     }