+++ src/app/components/forms/suggestion-box/example/terra-suggestion-box.component.example.ts
@@ -17,16 +17,20 @@ export class TerraSuggestionBoxComponentExample implements OnInit
 
     protected contacts:Array<any> = [
         {
+            name: 'Max',
+            age: 26
+        },
+        {
             name: 'Max Mustermann',
-            age: 28
+            age:  28
         },
         {
             name: 'Thomas Schmidt',
-            age: 28
+            age:  28
         },
         {
             name: 'Sabrina Meyer',
-            age: 29
+            age:  29
         }
     ];
     protected contactsSuggestions:Array<TerraSuggestionBoxValueInterface> = [];
@@ -39,31 +43,72 @@ export class TerraSuggestionBoxComponentExample implements OnInit
         this.iconList.push
         (
             {
-                value: 'icon-plugin',
+                value:   'icon-plugin',
                 caption: 'icon-plugin'
             },
             {
-                value: 'icon-delete',
+                value:   'icon-delete',
                 caption: 'icon-delete'
             },
             {
-                value:'icon-add',
+                value:   'icon-add',
                 caption: 'icon-add'
             },
             {
-                value: 'icon-box_plus',
+                value:   'icon-box_plus',
                 caption: 'icon-box_plus'
             },
             {
-                value: 'icon-flag_blue',
+                value:   'icon-flag_blue',
                 caption: 'icon-flag_blue'
+            },
+            {
+                value:   'icon-anonymize',
+                caption: 'icon-anonymize'
+            },
+            {
+                value:   'icon-active_doc',
+                caption: 'icon-active_doc'
+            },
+            {
+                value:   'icon-active_fold_open',
+                caption: 'icon-active_fold_open'
+            },
+            {
+                value:   'icon-real',
+                caption: 'icon-real'
+            },
+            {
+                value:   'icon-plenty_base',
+                caption: 'icon-plenty_base'
+            },
+            {
+                value:   'icon-servicecenter',
+                caption: 'icon-servicecenter'
+            },
+            {
+                value:   'icon-colorpicker',
+                caption: 'icon-colorpicker'
+            },
+            {
+                value:   'icon-align_left',
+                caption: 'icon-align_left'
+            },
+            {
+                value:   'icon-device_mobile',
+                caption: 'icon-device_mobile'
+            },
+            {
+                value:   'icon-ebay',
+                caption: 'icon-ebay'
             }
         );
 
-        this.contactsSuggestions = this.contacts.map(contact => {
+        this.contactsSuggestions = this.contacts.map((contact:any) =>
+        {
             return {
                 caption: contact.name,
-                value: contact
+                value:   contact
             };
         });
     }